[hofix] conflict 수정

+++ controllers/board.controller.js
@@ -1,5 +1,23 @@
 const boardService = require("../services/board.service");
-
+const geoInfo = [
+  "충청북도",
+  "강원도",
+  "경기도",
+  "경상남도",
+  "경상북도",
+  "광주광역시",
+  "대구광역시",
+  "대전광역시",
+  "부산광역시",
+  "서울특별시",
+  "세종특별시",
+  "울산광역시",
+  "인천광역시",
+  "전라남도",
+  "전라북도",
+  "제주특별자치도",
+  "충청남도",
+];
 class BoardController {
   async getBoards(req, res, next) {
     try {
@@ -20,6 +38,15 @@ class BoardController {
       next(err);
     }
   }
+  async getGeoStatus(req, res, next) {
+    try {
+      const result = await boardService.getGeoStatus();
+      console.log(result);
+      res.status(200).json(result);
+    } catch (err) {
+      next(err);
+    }
+  }
   async getBoard(req, res, next) {
     try {
       const id = req.params.id;
@@ -31,7 +58,10 @@ class BoardController {
   }
   async createBoard(req, res, next) {
     try {
-      const board = await boardService.createBoard(req.body);
+      const board = await boardService.createBoard({
+        ...req.body,
+        user_id: req.user.id,
+      });
       res.status(201).json(board);
     } catch (err) {
       next(err);
@@ -40,7 +70,12 @@ class BoardController {
   async updateBoard(req, res, next) {
     try {
       const id = req.params.id;
-      const board = await boardService.updateBoard(id, req.body);
+      if ((await boardService.getBoardById(id)) !== req.user.id)
+        throw new Error("permission denied");
+      const board = await boardService.updateBoard(id, {
+        ...req.body,
+        user_id: req.user.id,
+      });
       res.status(200).json(board);
     } catch (err) {
       next(err);
@@ -49,6 +84,8 @@ class BoardController {
   async deleteBoard(req, res, next) {
     try {
       const id = req.params.id;
+      if ((await boardService.getBoardById(id)) !== req.user.id)
+        throw new Error("permission denied");
       const board = await boardService.deleteBoard(id);
       res.status(200).json(board);
     } catch (err) {
